backend/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
channels==4.0.0

channels_redis

daphne[tls,http2]
Twisted[tls,http2]==22.10.0

psycopg2

Pillow>=8.0.0
pyotp

django_environ
requests
django-allauth
requests_oauthlib

gunicorn
whitenoise==6.0.0

djangorestframework-simplejwt
